learn about permission and group of permission

+++ src/main/java/com/discord/demo_app/model/Permission.java
@@ -9,7 +9,7 @@ import lombok.Data;
 public class Permission {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long permissionId;
+    private Integer permissionId;
     private String name;
     private String description;
 }
